pyproject.toml
[project]
name = "floor-quizzo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.92.0",
    "python-dotenv>=1.1.1",
    "st-gsheets-connection>=0.1.0",
    "streamlit>=1.46.1",
    "watchdog>=6.0.0",
]
